Attempted PVS again...
- Modified scale to be closer to source engines (gotten from icewinds
  renderer)
- Added config to configure some bits easily

I'm thinking that the next thing that I would need to do so is simply
get the bsp tree, and do something to visualize how it is choosing in
the hope that it would give me an idea of what was going wrong.
I think the root cause is something to do with leaf_at, though its
implementation is quite simple, so possibly some way I"m using the wrong
type of position that it would expect despite moving to the source
coordinate space and trying various mixups of that.